+++ internal/service/withdraw.go
@@ -8,7 +8,7 @@ type WithdrawService struct {
 	WithdrawRepository *repository.WithdrawRepository
 }
 
-func (ws *WithdrawService) Withdraw(userID int, orderNumber string, sum int) (int, error) {
+func (ws *WithdrawService) Withdraw(userID int, orderNumber string, sum float32) (int, error) {
 	return ws.WithdrawRepository.Withdraw(userID, orderNumber, sum)
 }
